+++ src/main/java/org/swisspush/redisques/queue/UnregisterConsumerType.java
@@ -1,0 +1,5 @@
+package org.swisspush.redisques.queue;
+
+public enum UnregisterConsumerType {
+    FORCE, GRACEFUL, QUIET_FOR_SOMETIME
+}
